Support Ambient Weather weather service.
The [Ambient Weather service](https://dashboard.ambientweather.net/) records
the weather data sent by an Ambient Weather weather station.  The API
documentation is at https://ambientweather.docs.apiary.io/.

The documentation says that API requests are limited to 3 requests per second
per application key, and that the API will return a 429 response if this rate
is exceeded.  If this becomes an issue, new application keys can be created
with a link on the bottom of the [Ambient Weather account
page](https://dashboard.ambientweather.net/account).