Persist current values
Persist the current values so that when HomeBridge is restarted it doesn't take up to a minute to get the values. Also this will ensure that the current Daily Rainfall total is close to being correct. Rain over the last hour is lost still during a restart.
Also adding Tempest/Weather Flow support  is a significant update to the plugin bump the version to 3.3