Remove Dark Sky support and fix an issue with Tempest and an issue with Open Weather Map. (#276)
* Remove DarkSky support #267

DarkSky is no longer available, remove support for it.

* Fix crash in #271

The error condition was not being handled. The function getWeatherData) can return an error and in that case, the variable 'result' is not valid.

* Make more use of converter

With the removal of DarkSky, update converter.getRainAccumulated to use floats and to sum up one array.
Move converter function for getWetBulbTemperature from Tempest to converter class. Then other weather stations can make use of function.

* Fix for #275

Initialise RainBool to false to fix warnings of undefined value.

* Fix issue identified in #276

Added the missing bracket, and corrected the function name declaration.
Also fixed the function indent to match the rest of the file.