Fix Rain Accumulation
Found an additional error now that it has been raining for a while. There is code to handle the case that if several minutes pass between observations, then clear out the old data in the skipped minutes. This was also clearing the last minute, so each time it clicked over a minute, it would erase the last one. This is fixed by clearing just the minutes that are actually skipped, one passed the last recorded to the current minute.