Fix forecasts to be error resilient
#285  Patch feedback
Fix crash when the apiKey and/or stationId is missing. Use the lastForecastUpdate variable to indicate if we have valid parameters at all. If they are not set, don't define the variable, and as such never attempt to fetch a forecast. This is a valid configuration so I don't log this state.
If invalid apiKey and/or stationId is used, then the URL request will fail with a response code other than 200 (OK). If that happens, fail with an error. This error will get logged. As the forecast is only fetched every hour, this error will only appear once an hour and won't spam the logs.