Forecast said there will be rain when there won't be
Patch feedback #285
Should not use the precipitation icons for snow and rain state as the chance precipitation can be zero, but there are still icons.
Instead do the same as OpenWeatherMap, normalize the conditions to the condition details numbered values, then check to see if there will be rain or snow.