Fix today's forecast
Today's forecast isn't sent after around 11pm.
To detect this, make sure that the current month day is the same as the month day of the forecast. If they are different, then offset the forecasts.
Save the current day forecast, so that when it is not available, the last forecast can be inserted into the forecasts array.